+++ src/torch_uncertainty/utils/__init__.py
@@ -5,7 +5,7 @@ from .cli import TULightningCLI
 from .distributions import NormalInverseGamma, get_dist_class, get_dist_estimate
 from .evaluation_loop import TUEvaluationLoop
 from .hub import load_hf
-from .misc import csv_writer, get_logger_dir, log_figure, log_image_array
+from .logging import csv_writer, get_logger_dir, log_figure, log_image_array
 from .plotting import plot_hist, show_segmentation_predictions
 from .trainer import TUTrainer
 from .transforms import interpolation_modes_from_str
